% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroGenixAssociation.R
\name{MicroGenixAssociation}
\alias{MicroGenixAssociation}
\title{Testing Associations between Phenotype and Predicted Host Gene Expression}
\usage{
MicroGenixAssociation(
  pred_expr,
  metadata,
  pheno_type = "binary",
  model = NULL,
  pheno = NULL,
  fixed_effect = NULL,
  random_effect = NULL,
  sig_thres = 0.05
)
}
\arguments{
\item{pred_expr}{a named numeric vector of predicted gene expression generated by MicroGenixPredict.}

\item{metadata}{the input file of metadata.}

\item{pheno_type}{type of phenotype. 'binary'(binary traits) or 'quant'(quantitative traits)}

\item{model}{the file name of a model object (.rds). Optional, if provided, interaction terms are returned.}

\item{pheno}{the column name of phenotype in the metadata file.}

\item{fixed_effect}{a character or character vector of the column names for the fixed effects (apart from pheno) in the metadata file.}

\item{random_effect}{a character of the column name for the random effects in the metadata file.}

\item{sig_thres}{significance threshold for the association. Interaction terms are not returned for insignificant associations.}
}
\value{
beta and p-value from the association test. Interaction terms for significant associations (if model is provided).
}
\description{
MicroGenixAssociation performs association tests between phenotype and gene expression levels predicted using MicroGenixPredict.
}
\examples{
\dontrun{
input_taxa <- system.file('extdata',
    'example_taxon_abundance.csv', package="MicroGenix")
input_geno <- system.file('extdata',
    'example_genotype_dosage.csv', package="MicroGenix")
model <- 'example_output_model.rds'
predicted_data <- MicroGenixPredict(model, input_taxa, input_geno)
pred_expr <- predicted_data$predicted_expr
metadata <- system.file('extdata',
    'example_metadata.csv', package="MicroGenix")
assoc_result <- MicroGenixAssociation(pred_expr, metadata,
    pheno = 'pheno', fixed_effect = 'cov_1')
}
}
