% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroGenixPredict.R
\name{MicroGenixPredict}
\alias{MicroGenixPredict}
\title{Predicting Host Gene Expression Using Models with Genotype-Microbiome Interactions}
\usage{
MicroGenixPredict(
  model,
  input_taxa,
  input_geno,
  observed_expr = NULL,
  log_trans_taxa = TRUE,
  log_trans_taxa_add = 0.01,
  log_trans_gene = FALSE,
  log_trans_gene_add = 0.01
)
}
\arguments{
\item{model}{the file name of a model object (.rds).}

\item{input_taxa}{the input file name of taxonomic abundance profiles.}

\item{input_geno}{the input file name of host genotype dosage (0/1/2).}

\item{observed_expr}{the file name of observed (actual) gene expression levels. If provided, accuracy is evaluated.}

\item{log_trans_taxa}{logical. If TRUE the taxonomic profiles are log2-transformed.}

\item{log_trans_taxa_add}{a numeric value of pseudo-count added to the taxonomic profiles to avoid taking the logarithm of zero.}

\item{log_trans_gene}{logical. If TRUE the observed gene expression levels are log2-transformed.}

\item{log_trans_gene_add}{a numeric value of pseudo-count added to the observed gene expression levels to avoid taking the logarithm of zero.}
}
\value{
Predicted gene expression levels and R2 (if observed_expr is provided).
}
\description{
MicroGenixPredict predicts gene expression using models trained with MicroGenixTrain.
}
\examples{
\dontrun{
input_taxa <- system.file('extdata',
    'example_taxon_abundance.csv', package="MicroGenix")
input_geno <- system.file('extdata',
    'example_genotype_dosage.csv', package="MicroGenix")
model <- 'example_output_model.rds'
predicted_data <- MicroGenixPredict(model, input_taxa, input_geno)
}
}
