% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroGenixTrain.R
\name{MicroGenixTrain}
\alias{MicroGenixTrain}
\title{Training Models with Genotype-Microbiome Interactions}
\usage{
MicroGenixTrain(
  input_taxa,
  input_geno,
  input_expr,
  log_trans_taxa = TRUE,
  log_trans_taxa_add = 0.01,
  log_trans_gene = FALSE,
  log_trans_gene_add = 0.01,
  fold_id = NULL,
  num_folds = 10,
  output_prefix = NULL,
  pval_fit = FALSE
)
}
\arguments{
\item{input_taxa}{the input file name of taxonomic abundance profiles.}

\item{input_geno}{the input file name of host genotype dosage (0/1/2).}

\item{input_expr}{the input file name of host gene expression levels.}

\item{log_trans_taxa}{logical. If TRUE the taxonomic profiles are log2-transformed.}

\item{log_trans_taxa_add}{a numeric value of pseudo-count added to the taxonomic profiles to avoid taking the logarithm of zero.}

\item{log_trans_gene}{logical. If TRUE the gene expression levels are log2-transformed.}

\item{log_trans_gene_add}{a numeric value of pseudo-count added to the gene expression levels to avoid taking the logarithm of zero.}

\item{fold_id}{"foldid" for cv.glmnet, see \code{\link[glmnet]{cv.glmnet}}}

\item{num_folds}{"nfolds" for cv.glmnet, see \code{\link[glmnet]{cv.glmnet}}}

\item{output_prefix}{the prefix of output file names.}

\item{pval_fit}{logical. This is an experimental argument. If TRUE fit a basic linear model using variables selected by elastic net to obtain p-values.}
}
\value{
A model object and coefficients.
}
\description{
MicroGenixTrain is used to fit models to predict gene expression with elastic net. Predictors include SNPs, microbes, and SNP-microbe interactions.
}
\examples{
\dontrun{
input_taxa <- system.file('extdata', 'example_taxon_abundance.csv', package="MicroGenix")
input_geno <- system.file('extdata','example_genotype_dosage.csv', package="MicroGenix")
input_expr <- system.file('extdata', 'example_gene_expression.csv', package="MicroGenix")
fit_data <- MicroGenixTrain(input_taxa, input_geno, input_expr)
}
}
